pyproject.toml
[project]
name = "uwazi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "cloudinary>=1.44.1",
    "fastapi[standard]>=0.128.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "prophet>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.46",
    "xgboost>=3.2.0",
]
